refactor: Move theme to UI
