+++ ui/header/index.ts
@@ -1,2 +1,3 @@
 export * from "./header";
 export * from "./menu";
+export * from "./theme";
